@@ -5,7 +5,7 @@ author: Max<vhbvbqc@gmail.com>
 homepage: https://www.mob.com/product/sharesdk
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
